requirements.txt
torch>=2.1.0
ultralytics>=8.0.0
opencv-python>=4.7.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
